Move some apply transaction class from diagram package to apply_transaction package
